Small refactoring.
