+++ src/vs15/CoCo.Analyser/QuickInfo/SymbolDescriptionProvider.XmlDocumentParser.cs
