+++ src/vs15/CoCo.Analyser/VisualBasic/VisualBasicClassifier.cs
@@ -40,7 +40,7 @@ namespace CoCo.Analyser.VisualBasic
 
         internal VisualBasicClassifier(
             IReadOnlyDictionary<string, ClassificationInfo> classifications,
-            IAnalyzingService analyzingService,
+            IClassificationChangingService analyzingService,
             ITextDocumentFactoryService textDocumentFactoryService,
             ITextBuffer buffer) : base(analyzingService, textDocumentFactoryService, buffer)
         {
