+++ src/vs15/CoCo/Providers/VisualBasicClassifierProvider.cs
@@ -33,7 +33,7 @@ namespace CoCo.Providers
             {
                 _classificationsInfo[item] = default;
             }
-            AnalyzingService.Instance.ClassificationChanged += OnAnalyzeOptionChanged;
+            ClassificationChangingService.Instance.ClassificationChanged += OnAnalyzeOptionChanged;
         }
 
 #pragma warning disable 649
@@ -54,12 +54,12 @@ namespace CoCo.Providers
                 var settings = SettingsManager.LoadEditorSettings(Paths.CoCoSettingsFile, MigrationService.Instance);
                 var option = OptionService.ToOption(settings);
                 FormattingService.SetFormattingOptions(option);
-                AnalyzingService.SetAnalyzingOptions(option);
+                ClassificationChangingService.SetAnalyzingOptions(option);
                 _wereSettingsSet = true;
             }
 
             return textBuffer.Properties.GetOrCreateSingletonProperty(() =>
-                new VisualBasicClassifier(_classificationsInfo, AnalyzingService.Instance, _textDocumentFactoryService, textBuffer));
+                new VisualBasicClassifier(_classificationsInfo, ClassificationChangingService.Instance, _textDocumentFactoryService, textBuffer));
         }
 
         private void OnAnalyzeOptionChanged(ClassificationsChangedEventArgs args)
